Release ndk 0.4.0, ndk-glue 0.4.0, ndk-build 0.4.1 (#166)
ndk-glue depends on a breaking change in the NDK, and can only be
released if that's released under 0.4.0.  This, in turn, causes ndk-glue
to need a minor bump too despite no other breaking changes, as NDK
symbols are used in its public API.

Note that ndk-glue 0.3.1 was never properly released as the ndk update
requirement outlined above caused an adequate publishing failure.