ndk-build: Move default serialization of `MAIN` intent filter to `cargo-apk` (#241)
The other two manual `serialize_with` functions have a very clear
purpose of making the datastructure easier to work with, by turning a
vector of strings into a more verbose XML element with the string as
`name` field.

No such case with `intent_filter`, which needs no special serialization
treatment but instead performs actual business logic to insert more
elements based on what the caller wishes to serialize.  This
sort of logic is better moved to `cargo-apk` where a toml-parsed
manifest is similarly enriched with sensible defaults to make Android
development as accessible and out-of-the-box as possible.