ndk-build: Make platform SDK parsing more robust and ignore minor suffix
The minor suffix of an SDK version is not something that can be targeted
or conveyed in an Android app; ignore it for now as we only use this
to determine the highest API level to target.

However `fn platform_dir()` independently tries to reconstruct this
path based on an API target and now has no idea to append `.1` if only
`android-36.1` happens to be installed for instance.  This should be
brought back in sync.